feat: implementa melhoria de estrutura em DTOs
