--- backend/DTOs/EncomendaCreateDto.cs
@@ -1,12 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-
-namespace backend.DTOs;
-
-public class EncomendaCreateDto
-{
-    [Required]
-    public string Remetente { get; set; } = string.Empty;
-
-    [Required]
-    public long MoradorId { get; set; }
-}
