+++ backend/DTOs/EncomendaDtos.cs
@@ -1,6 +1,16 @@
+using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
 
 namespace backend.DTOs;
+
+public class EncomendaCreateDto
+{
+    [Required]
+    public string Remetente { get; set; } = string.Empty;
+
+    [Required]
+    public long MoradorId { get; set; }
+}
 
 public class EncomendaResponseDto
 {
@@ -17,4 +27,20 @@ public class EncomendaResponseDto
 
     public DateTime? DataRetirado { get; set; }
     public string Status { get; set; } = string.Empty;
+}
+
+public class EncomendaRetiradaDto
+{
+    [Required(ErrorMessage = "A informação de retirada é obrigatória.")]
+    public bool? Retirada { get; set; }
+}
+
+public class EncomendaUpdateDto
+{
+    [Required(ErrorMessage = "O status da encomenda é obrigatório.")]
+    public long IdCategoriaEncomenda { get; set; }
+
+    public DateTime? DataRetirado { get; set; }
+    
+    public string? Remetente { get; set; }
 }
