--- backend/DTOs/FuncionarioResponseDTO.cs
@@ -1,7 +1,0 @@
-namespace backend.DTOs;
-
-public class FuncionarioResponseDto
-{
-    public long CargoId { get; set; }
-    public string CargoNome { get; set; } = string.Empty;
-}
