+++ backend/DTOs/LoginDtos.cs
@@ -1,5 +1,19 @@
+using System.ComponentModel.DataAnnotations;
+
 namespace backend.DTOs;
 
+public class LoginDto
+{
+    /// <summary>CPF cadastrado (somente números).</summary>
+    /// <example>12345678901</example>
+    [Required, StringLength(11)]
+    public string Cpf { get; set; } = string.Empty;
+
+    /// <summary>Senha definida no cadastro.</summary>
+    /// <example>Senha@123</example>
+    [Required]
+    public string Senha { get; set; } = string.Empty;
+}
 public class LoginResponseDTO
 {
     /// <summary>Mensagem de status da operação.</summary>
