--- backend/DTOs/MoradorResponseDTO.cs
@@ -1,7 +1,0 @@
-namespace backend.DTOs;
-
-public class MoradorResponseDto
-{
-    public string Bloco { get; set; } = string.Empty;
-    public int Apartamento { get; set; }
-}
