--- backend/DTOs/ReclamacaoAdminUpdateDto.cs
@@ -1,9 +1,0 @@
-// DTOs/ReclamacaoAdminUpdateDto.cs
-using System.ComponentModel.DataAnnotations;
-
-namespace backend.DTOs;
-public class ReclamacaoAdminUpdateDto : ReclamacaoCreateDto
-{
-    [Required]
-    public long IdCategoriaReclamacao { get; set; }
-}
