--- backend/DTOs/ReclamacaoPublicaDto.cs
@@ -1,9 +1,0 @@
-namespace backend.DTOs;
-
-public class ReclamacaoPublicaDto
-{
-    public long Id { get; set; }
-    public string Titulo { get; set; } = string.Empty;
-    public string Descricao { get; set; } = string.Empty;
-    public string Status { get; set; } = string.Empty;
-}
