--- backend/DTOs/ReclamacaoUpdateDto.cs
@@ -1,6 +1,0 @@
-// DTOs/ReclamacaoUpdateDto.cs
-// O morador só pode alterar o texto, não o status.
-using System.ComponentModel.DataAnnotations;
-
-namespace backend.DTOs;
-public class ReclamacaoUpdateDto : ReclamacaoCreateDto { }
