--- backend/DTOs/ReservaResponseDTO.cs
@@ -1,12 +1,0 @@
-namespace backend.DTOs;
-
-public class ReservaResponseDto
-{
-    public long Id { get; set; }
-    public long IdLocal { get; set; }
-    public DateTime DataInicio { get; set; }
-    public DateTime DataFim { get; set; }
-    public string Status { get; set; } = string.Empty;
-
-
-}
