--- backend/DTOs/ReservaUpdateDto.cs
@@ -1,15 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-
-namespace backend.DTOs;
-
-public class ReservaUpdateDto
-{
-    [Required(ErrorMessage = "O local da reserva é obrigatório.")]
-    public long IdLocal { get; set; }
-
-    [Required(ErrorMessage = "A data de início é obrigatória.")]
-    public DateTime DataInicio { get; set; }
-
-    [Required(ErrorMessage = "A data de término é obrigatória.")]
-    public DateTime DataFim { get; set; }
-}
